+++ lib/express_templates/components/forms.rb
@@ -6,5 +6,7 @@ module ExpressTemplates
 end
 
 require 'express_templates/components/forms/express_form'
+require 'express_templates/components/forms/form_component'
+require 'express_templates/components/forms/selectable'
 require 'express_templates/components/forms/submit'
 require 'express_templates/components/forms/select'
